Add the ability to remove modifications.
Rename prune to prune_confs and add prune_atoms, which removes any atoms not present in the unmodified residue, the simplest case of undoing a posttranslational modification. Add new lambdas for this purpose to each ptm dictionary item and list the atoms in the unmodified residue in the unmodified residue dictionary items.