Add functionality for testing on synthetic data generated from a model.
When requested, test on synthetic data instead of an experimental map. Supply a model and map (used only for sizing the synthetic map). First, copy the model and add posttranslational modifications at random to 10% of recognized residues. Calculate a map from this model and substitute it for the experimental map. Use the unmodified model as the baseline and searh for modifications as usual. Then compare identified and placed modifications to test success rates.